remove all dependencies between wasi-common and wasmtime-wasi
